requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4